Type casting fixes

+++ plugins/Oculars/src/gui/OcularsGuiPanel.cpp
@@ -174,7 +174,7 @@ OcularsGuiPanel::OcularsGuiPanel(Oculars* plugin,
 	float scv = plugin->getSettings()->value("arrow_scale", 150.f).toFloat();
 	if (scv<100.f) // convert old value and type
 		scv *= 100.f;
-	float scale=lineHeight*scv*0.01f;
+	int scale=static_cast<int>(lineHeight*scv*0.01f);
 	// TODO: change this load-once to interactively editable value of scaling coefficient
 	QPixmap pa(":/graphicGui/btTimeRewind-on.png");
 	QPixmap prevArrow = pa.scaledToHeight(scale, Qt::SmoothTransformation);
